Normalize units in ZFS properties from playbooks
`zfs set` treats units provided to it for properties such as 'quota'
and 'reservation' in a case-insensitive manner. Additionaly, it will
also accept an optional 'b' (for bytes). That means M, mb, mB, etc.
are all valid abbreviations for megabytes.

`zfs get` always returns the unit as a single uppercase letter. When the
zfs module checks to see if it needs to update any properties, it always
compares what's provided in the playbook with what's returned by `zfs
get`. As a result, if a playbook specifies units in a format other than a
single uppercase letter, that property will always show as changed and
Ansible will attempt to update it to the correct value.

This commit adds code to normalize units that the `zfs set` command
would accept to the single uppercase letter format before comparing it
to the output of `zfs get`.